eslint-config-target-es: Improve use of MDN data (#31658)
Preeviously we had been doing the bare minimum: we only looked at the
`version_added` field for the most recent support statement for each
feature, compared against the lowest version for each browser returned
from browserslist. Mostly that works, but there are some edge cases it
could miss (e.g. multiple active branches of a browser where something
was added in 9.6 in the 9.x branch and 10.2 in the 10.x branch).

Also, while we had been considering `partial_implementation` as
indicating that the rule should still be enabled, we weren't doing the
same for `flags` or `prefix` or `alternative_name` which are even more
likely to indicate a potential lack of support.

With this PR we now look at all the support statements, consider
`version_removed` in each, and take note of `flags` and `prefix` and
`alternative_name` when those are present.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/5444371136